+++ helm/client_test.go
@@ -1,7 +1,6 @@
 package helm
 
 import (
-	"bytes"
 	"context"
 	"io/ioutil"
 	"testing"
@@ -11,13 +10,11 @@ import (
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/chart"
 	"helm.sh/helm/v3/pkg/chartutil"
-	"helm.sh/helm/v3/pkg/kube"
 	"helm.sh/helm/v3/pkg/kube/fake"
 	"helm.sh/helm/v3/pkg/release"
 	"helm.sh/helm/v3/pkg/storage"
 	"helm.sh/helm/v3/pkg/storage/driver"
 	"helm.sh/helm/v3/pkg/time"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	"github.com/castai/cluster-controller/castai"
 )
@@ -96,96 +93,6 @@ func TestClientUninstall(t *testing.T) {
 		Namespace:   currentRelease.Namespace,
 	})
 	r.NoError(err)
-}
-
-func TestIgnoreHook(t *testing.T) {
-	r := require.New(t)
-
-	oldManifests :=
-		`---
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  labels:
-    app.kubernetes.io/instance: castai-evictor
-    app.kubernetes.io/managed-by: Helm
-    app.kubernetes.io/name: castai-evictor
-    app.kubernetes.io/version: 0.5.1
-    helm.sh/chart: castai-evictor-0.10.0
-  name: castai-evictor
-
----
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    app.kubernetes.io/instance: castai-evictor
-    app.kubernetes.io/managed-by: Helm
-    app.kubernetes.io/name: castai-evictor
-    app.kubernetes.io/version: 0.5.1
-    helm.sh/chart: castai-evictor-0.10.0
-  name: castai-evictor
-  namespace: castai-agent`
-
-	newManifests :=
-		`---
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  labels:
-    app.kubernetes.io/instance: castai-evictor
-    app.kubernetes.io/managed-by: Helm
-    app.kubernetes.io/name: castai-evictor
-    app.kubernetes.io/version: 0.6.0
-    helm.sh/chart: castai-evictor-0.11.0
-  name: castai-evictor
-
----
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    app.kubernetes.io/instance: castai-evictor
-    app.kubernetes.io/managed-by: Helm
-    app.kubernetes.io/name: castai-evictor
-    app.kubernetes.io/version: 0.6.0
-    helm.sh/chart: castai-evictor-0.11.0
-  name: castai-evictor
-  namespace: castai-agent`
-
-
-	oldRelease := &release.Release{
-		Manifest: oldManifests,
-	}
-
-	cl := kube.New(nil)
-
-	hook := LabelIgnoreHook{
-		oldRelease: oldRelease,
-		kubeClient: cl,
-	}
-
-	buf := bytes.NewBuffer([]byte(newManifests))
-
-	fixedManifest, err := hook.Run(buf)
-	r.NoError(err)
-
-	typed, err := cl.Build(fixedManifest, false)
-	r.NoError(err)
-
-	for _, res := range typed {
-		u := res.Object.(*unstructured.Unstructured)
-
-		if u.GetKind() == "Service" {
-			r.Equal("0.6.0", u.GetLabels()[k8sVersionLabel])
-			r.Equal("castai-evictor-0.11.0", u.GetLabels()[helmVersionLabel])
-		}
-
-		if u.GetKind() == "ClusterRoleBinding" {
-			r.Equal("0.5.1", u.GetLabels()[k8sVersionLabel])
-			r.Equal("castai-evictor-0.10.0", u.GetLabels()[helmVersionLabel])
-		}
-	}
 }
 
 type testConfigurationGetter struct {
